+++ gbif_blocking_occurrences_download.py
@@ -1,4 +1,5 @@
 import logging
+import pathlib
 from time import sleep
 
 import requests
@@ -8,15 +9,29 @@ def download_occurrences(
     predicate: dict,
     username: str,
     password: str,
-    output_path: str,
+    output_path: pathlib.Path,
     sleep_duration: int = 20,
 ) -> None:
+    """
+    Request an occurrence download at GBIF, wait for it to be ready and finally save the result.
+
+    :param predicate: a predicate to select included occurrences. See: https://www.gbif.org/developer/occurrence#predicates
+    :param username: a valid username (on GBIF.org)
+    :param password: a valid password (on GBIF.org)
+    :param output_path: a path and filename where the download will be saved
+    :param sleep_duration: when waiting for the results to be ready, how long to wait between retries
+    :return: nothing
+    :raise: requests.exceptions.HTTPError when the occurrence download cannot be requested (i.e. 401 if credentials are invalid)
+    """
 
     download_trigger = requests.post(
         "https://api.gbif.org/v1/occurrence/download/request",
         json=predicate,
         auth=(username, password),
     )
+
+    download_trigger.raise_for_status()
+
     download_id = download_trigger.text
     logging.info(f"Download triggered, download_id is {download_id}")