+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v0.1.1 (2021-08-27)
+===================
+
+* An exception is now raised if the credentials are incorrect
+* Basic test suite, docstrings, ...
+
 v0.1.0 (2021-08-26)
 ===================